requirements.txt
streamlit>=1.34
opencv-contrib-python-headless>=4.10
numpy>=1.26
Pillow>=10.3
